+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 2024.11.11 (2024-11-25)
 
 - Enable central link management for HmIP-wired
+- Reset temporary values before write
 
 # Version 2024.11.10 (2024-11-24)